Replace make indicators compute in double

+++ server/feature/portfolios/src/main/java/ru/kima/sonar/server/feature/portfolios/service/model/CacheEntry.kt
@@ -1,23 +1,22 @@
 package ru.kima.sonar.server.feature.portfolios.service.model
 
 import ru.kima.sonar.server.feature.portfolios.techanalysis.BollingerBands
-import java.math.BigDecimal
 
 data class CacheEntry(
-    val min15Rsi: BigDecimal,
-    val hourlyRsi: BigDecimal,
-    val hour4Rsi: BigDecimal,
-    val dailyRsi: BigDecimal,
+    val min15Rsi: Double,
+    val hourlyRsi: Double,
+    val hour4Rsi: Double,
+    val dailyRsi: Double,
     val min15bb: BollingerBands.BollingerBandsData,
     val hourlyBb: BollingerBands.BollingerBandsData,
     val hour4Bb: BollingerBands.BollingerBandsData,
     val dailyBb: BollingerBands.BollingerBandsData,
-    val min15Mfi: BigDecimal,
-    val hourlyMfi: BigDecimal,
-    val hour4Mfi: BigDecimal,
-    val dailyMfi: BigDecimal,
-    val min15Srsi: BigDecimal,
-    val hourlySrsi: BigDecimal,
-    val hour4Srsi: BigDecimal,
-    val dailySrsi: BigDecimal,
+    val min15Mfi: Double,
+    val hourlyMfi: Double,
+    val hour4Mfi: Double,
+    val dailyMfi: Double,
+    val min15Srsi: Double,
+    val hourlySrsi: Double,
+    val hour4Srsi: Double,
+    val dailySrsi: Double,
 )
